test(core,widgets): move layout protection test to widgets package to avoid circular dependency
